Messy commit doing to much at once.
Move DisplayFile and DisplayMetadata to a new home.
Add a super class DisplatItem.
Use this Class in a new function display_item at the project level.
Use this new functionality in the project browser to display files.
